% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backends}
\alias{backends}
\title{List registered backends}
\usage{
backends()
}
\value{
A data frame with one row per backend: \code{name}, \code{official}, \code{description}.
}
\description{
List registered backends
}
\examples{
backends()
}
